Attempt to fix some frame rebuilding issues
Attempt to fix issues where member flag changes rebuilded all the frames
and caused noticable issues when fade when out-of-range option was used
since the opacity was reset for a second. Now only rebuilds the frames
when absolutely required, this gets rid of most of those issues. But it
still occurs when changing roles/becoming leader and such, since that
requires frame rebuilding.